requirements.txt
alembic==0.9.9
argh==0.26.2
click==6.7
Flask==1.0.2
Flask-API==1.0
Flask-Cors==3.0.4
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
nose==1.3.7
nose-watch==0.9.2
pathtools==0.1.2
python-dateutil==2.7.3
python-editor==1.0.3
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.7
watchdog==0.8.3
Werkzeug==0.14.1
